refactor: src 레이아웃 적용으로 프로젝트 구조 변경
