+++ src/camera_stream/kafka_streams.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from config.kafka_broker_instance import broker
+from src.config.kafka_broker_instance import broker
 
 async def connect_broker():
     try:
